requirements.txt
# Web Framework
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2

# Template Engine
Jinja2==3.1.6
MarkupSafe==3.0.2

# Database - MongoDB
beanie==1.26.0
motor==3.6.0

# Authentication & Security
passlib==1.7.4
bcrypt==3.2.2
python-jose==3.5.0
cryptography==45.0.5
python-multipart==0.0.20

# Validation
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.7.0
email-validator==2.3.0

# HTTP Client
httpx==0.28.1
httpcore==1.0.9

# Testing
pytest==8.4.1
pytest-asyncio==1.1.0

# Utilities
aiofiles==24.1.0
click==8.1.8
APScheduler==3.10.4

# Dependencies
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
cffi==1.17.1
ecdsa==0.19.1
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
Pygments==2.19.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1

# Deprecated - kept for compatibility
asyncio==3.4.3
backports.asyncio.runner==1.2.0

